changing harvester spawning function to use room energy capacity instead of available
this is to counter the problem where many creeps dying at once meant
that the extensions couldn't be refilled fast enough and the creeps
would end up being undersize because the spawn wouldn't wait for more
energy. this would happen continuously because they would die and
respawn in the same order close together every time.